generated_code/fixtures/tests/test_commonloggingstatements_fixture.py
from code_autoeval.llm_model.utils.base_llm_class import BaseLLMClass
from code_autoeval.llm_model.utils.base_llm_class import BaseLLMClass
import pytest
from generated_code.fixtures.fixtures.commonloggingstatements_fixture import fixture_mock_commonloggingstatements
from code_autoeval.llm_model.utils.logging_statements.common_logging_statements import CommonLoggingStatements
from pathlib import Path
from typing import Any, Dict
def test_mock_commonloggingstatements(fixture_mock_commonloggingstatements):
    assert isinstance(fixture_mock_commonloggingstatements, CommonLoggingStatements)
    assert hasattr(fixture_mock_commonloggingstatements, '_log_code')
    assert callable(fixture_mock_commonloggingstatements._log_code)
    assert hasattr(fixture_mock_commonloggingstatements, '_log_coverage_results')
    assert callable(fixture_mock_commonloggingstatements._log_coverage_results)
    assert hasattr(fixture_mock_commonloggingstatements, '_log_fake_gen_data')
    assert callable(fixture_mock_commonloggingstatements._log_fake_gen_data)
    assert hasattr(fixture_mock_commonloggingstatements, '_log_max_retries')
    assert callable(fixture_mock_commonloggingstatements._log_max_retries)
    assert hasattr(fixture_mock_commonloggingstatements, '_log_test_coverage_path')
    assert callable(fixture_mock_commonloggingstatements._log_test_coverage_path)
    assert hasattr(fixture_mock_commonloggingstatements, 'coverage_result')
    assert hasattr(fixture_mock_commonloggingstatements, 'file_path')
    assert isinstance(fixture_mock_commonloggingstatements.file_path, Path)
    assert hasattr(fixture_mock_commonloggingstatements, 'test_file_path')
    assert isinstance(fixture_mock_commonloggingstatements.test_file_path, Path)
    assert hasattr(fixture_mock_commonloggingstatements, 'absolute_path_from_root')
    assert isinstance(fixture_mock_commonloggingstatements.absolute_path_from_root, Path)
    assert hasattr(fixture_mock_commonloggingstatements, 'unique_imports_dict')
    assert isinstance(fixture_mock_commonloggingstatements.unique_imports_dict, Dict)
    assert isinstance(fixture_mock_commonloggingstatements, BaseLLMClass)